Break up mega-implementation into multiple smaller files
