+++ runtime/serde/serde-cbor/common/test/aws/smithy/kotlin/runtime/serde/cbor/CborDeserializerSuccessTest.kt
@@ -7,6 +7,7 @@ package aws.smithy.kotlin.runtime.serde.cbor
 import aws.smithy.kotlin.runtime.io.SdkBuffer
 import aws.smithy.kotlin.runtime.serde.SdkFieldDescriptor
 import aws.smithy.kotlin.runtime.serde.SerialKind
+import aws.smithy.kotlin.runtime.serde.cbor.encoding.NegInt
 import aws.smithy.kotlin.runtime.serde.deserializeList
 import aws.smithy.kotlin.runtime.serde.deserializeMap
 import kotlin.test.*
@@ -192,7 +193,7 @@ class CborDeserializerSuccessTest {
     fun `atomic - negint - 8 - max`() {
         val payload = "0x3bfffffffffffffffe".toByteArray()
         val buffer = SdkBuffer().apply { write(payload) }
-        val result = Cbor.Encoding.NegInt.decode(buffer).value
+        val result = NegInt.decode(buffer).value
         assertEquals(ULong.MAX_VALUE, result)
     }
 
@@ -322,7 +323,7 @@ class CborDeserializerSuccessTest {
     fun `atomic - negint - 2 - max`() {
         val payload = "0x39ffff".toByteArray()
         val buffer = SdkBuffer().apply { write(payload) }
-        val result = Cbor.Encoding.NegInt.decode(buffer).value
+        val result = NegInt.decode(buffer).value
         assertEquals(65536u, result)
     }
 
@@ -1188,7 +1189,7 @@ class CborDeserializerSuccessTest {
 
         val remainingBuffer = "0x3bfffffffffffffffe".toByteArray()
         val buffer = SdkBuffer().apply { write(remainingBuffer) }
-        val result = Cbor.Encoding.NegInt.decode(buffer).value
+        val result = NegInt.decode(buffer).value
         assertEquals(ULong.MAX_VALUE, result)
     }
 
@@ -1497,7 +1498,7 @@ class CborDeserializerSuccessTest {
 
         val remainingBuffer = "0x3bfffffffffffffffe".toByteArray()
         val buffer = SdkBuffer().apply { write(remainingBuffer) }
-        val result = Cbor.Encoding.NegInt.decode(buffer).value
+        val result = NegInt.decode(buffer).value
         assertEquals(ULong.MAX_VALUE, result)
     }
 
@@ -1781,7 +1782,7 @@ class CborDeserializerSuccessTest {
 
         val remainingBuffer = "0x3bfffffffffffffffe".toByteArray()
         val buffer = SdkBuffer().apply { write(remainingBuffer) }
-        val result = Cbor.Encoding.NegInt.decode(buffer).value
+        val result = NegInt.decode(buffer).value
         assertEquals(ULong.MAX_VALUE, result)
     }
 
@@ -2235,7 +2236,7 @@ class CborDeserializerSuccessTest {
 
         val remainingBuffer = "0x3bfffffffffffffffe".toByteArray()
         val buffer = SdkBuffer().apply { write(remainingBuffer) }
-        val result = Cbor.Encoding.NegInt.decode(buffer).value
+        val result = NegInt.decode(buffer).value
         assertEquals(ULong.MAX_VALUE, result)
     }
